add individual Delivery and Billing

+++ app/View/Elements/backend/portlets/Billings/tableContent.ctp
@@ -1,7 +1,7 @@
 <?php	
 	foreach ($data as $item):
 	
-		if($item['Confirmation']['cart_id'] != 0) {
+		if(true) {
 			
 		$datetime1 = new DateTime();
 		$datetime2 = new DateTime($item['Billing']['payment_target']);
@@ -17,34 +17,44 @@
 						<?php if(empty($item['Billing']['billing_number'])) {
 							 echo '-'; 
 							} else {
-								echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'view', $item['Billing']['id']), array('escape' => false));
-								echo '&nbsp;&nbsp;&nbsp;';
-								echo $item['Billing']['billing_number'];	
+								
+								if(strpos($item['Confirmation']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'edit_individual', $item['Billing']['id']), array('escape' => false));
+									echo '&nbsp;&nbsp;&nbsp;';
+									echo $item['Billing']['billing_number'];
+								} else {
+									echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'view', $item['Billing']['id']), array('escape' => false));
+									echo '&nbsp;&nbsp;&nbsp;';
+									echo $item['Billing']['billing_number'];
+								}
+								
+								
+									
 							
 							}
 						?>
 					</td>
 					<td>
-						<?php 
-												
-						if(!is_null($item['Customer'])) {
+						<?php 						
+						if(!is_null($item['Confirmation']['customer_id'])) {
 							
 							if(empty($item['Confirmation']['customer_id'])) { echo '-'; } else {
-																	
 								echo $item['Confirmation']['customer_id'];	
-								echo '&nbsp;';
-								echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
-									 data-toggle="popover"
-									 data-content="';
-									 	echo $item['Address']['organisation'].'<br>';
-										if(!empty($item['Address']['department'])) {
-											echo $item['Address']['department'].'<br>';
-										}
-									 	echo $item['Address']['name'].'<br>'.
-									 '"
-									 data-trigger="hover"
-								
-								></i>';
+								if(!is_null($item['Billing']['address_id']) && $item['Billing']['address_id'] != 0) {
+									echo '&nbsp;';
+									echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
+										 data-toggle="popover"
+										 data-content="';
+										 	echo $item['Address']['organisation'].'<br>';
+											if(!empty($item['Address']['department'])) {
+												echo $item['Address']['department'].'<br>';
+											}
+										 	echo $item['Address']['name'].'<br>'.
+										 '"
+										 data-trigger="hover"
+									
+									></i>';
+								}
 							}
 						} else { echo '-'; } 
 						 ?>
@@ -52,19 +62,29 @@
 					<td>
 						<?php 
 						$cartProducts = "";
-						foreach ($item['Cart']['CartProduct'] as $cartProduct) {
-							$cartProducts = $cartProducts . $cartProduct['amount'].'x '. $cartProduct['Information']['Product']['name'] . ' ( '.$cartProduct['Information']['Product']['product_number'].' )<br>';
-						}					
-
-						$priceInfo = 'Gesamtpreis: '.$this->Number->currency($item['Cart']['sum_retail_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
-									 $item['Confirmation']['discount'].'% Rabatt: '.$this->Number->currency($item['Billing']['discount_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
-									 'Versandkostenvorteil: '.$this->Number->currency($item['Billing']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
-									 'Zwischensumme: '.$this->Number->currency($item['Billing']['part_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
-									 '+'.$item['Confirmation']['vat'].'% Mehrwertsteuer: '.$this->Number->currency($item['Billing']['vat_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
-									 'Rechnungswert: '.$this->Number->currency($item['Billing']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ','));					
 						
 	
-						if(empty($item['Cart']['count'])) { echo '-'; } else {
+						if(empty($item['Cart']['count'])) {
+							 
+							if(strpos($item['Billing']['status'], 'custom') !== FALSE){
+								echo $this->Number->currency($item['Confirmation']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ','));	
+							} else {
+								echo '-'; 							
+							}
+						} else {
+							
+							foreach ($item['Cart']['CartProduct'] as $cartProduct) {
+								$cartProducts = $cartProducts . $cartProduct['amount'].'x '. $cartProduct['Information']['Product']['name'] . ' ( '.$cartProduct['Information']['Product']['product_number'].' )<br>';
+							}					
+	
+							$priceInfo = 'Gesamtpreis: '.$this->Number->currency($item['Cart']['sum_retail_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
+										 $item['Confirmation']['discount'].'% Rabatt: '.$this->Number->currency($item['Billing']['discount_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
+										 'Versandkostenvorteil: '.$this->Number->currency($item['Billing']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
+										 'Zwischensumme: '.$this->Number->currency($item['Billing']['part_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
+										 '+'.$item['Confirmation']['vat'].'% Mehrwertsteuer: '.$this->Number->currency($item['Billing']['vat_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ',')).'<br>'.
+										 'Rechnungswert: '.$this->Number->currency($item['Billing']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ','));					
+							
+							
 							echo $this->Number->currency($item['Billing']['confirmation_price'],'EUR', array('wholePosition' => 'after', 'before' => ' €', 'thousands' => '.', 'decimals' => ','));	
 							echo '&nbsp;';
 							echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
@@ -127,14 +147,22 @@
 					<td>
 						<?php 
 							
-							if(empty($item['Cart']['count']) || $item['Confirmation']['order_date'] == '0000-00-00' || empty($item['Confirmation']['customer_id']) || empty($item['Confirmation']['confirmation_price'])) {
+							if($item['Confirmation']['order_date'] == '0000-00-00' || empty($item['Confirmation']['customer_id']) || empty($item['Confirmation']['confirmation_price'])) {
 								echo '-';
 							} else {
-								 echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
-								 data-content="'.$item['Confirmation']['confirmation_number'].'"
-								 data-trigger="hover"></i>', 
-								 array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));
 								
+								if(strpos($item['Confirmation']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									 data-content="'.$item['Confirmation']['confirmation_number'].'"
+									 data-trigger="hover"></i>', 
+									 array('admin' => true, 'controller' => 'Confirmations', 'action' => 'edit_individual', $item['Confirmation']['id']), array('escape' => false));
+								} else {
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									 data-content="'.$item['Confirmation']['confirmation_number'].'"
+									 data-trigger="hover"></i>', 
+									 array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));
+								}
+																	
 							}
 							  
 						 ?>
@@ -142,14 +170,24 @@
 					<!-- Lieferschein-Nummer -->
 					<td>
 						<?php 
-							if(empty($item['Cart']['count']) || $item['Confirmation']['order_date'] == '0000-00-00' || empty($item['Confirmation']['customer_id']) || empty($item['Confirmation']['confirmation_price'])) {
+							if($item['Confirmation']['order_date'] == '0000-00-00' || empty($item['Confirmation']['customer_id']) || empty($item['Confirmation']['confirmation_price'])) {
 								echo '-';
 							} else {
-								 echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
-								 data-content="'.$item['Billing']['delivery_number'].'"
-								 data-trigger="hover"
-							
-							></i>', array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));
+								
+								if(strpos($item['Confirmation']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									data-content="'.$item['Billing']['delivery_number'].'"
+									data-trigger="hover"
+									></i>', array('admin' => true, 'controller' => 'Confirmations', 'action' => 'edit_individual', $item['Confirmation']['id']), array('escape' => false));
+								
+								} else {
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									data-content="'.$item['Billing']['delivery_number'].'"
+									data-trigger="hover"
+									></i>', array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));
+								
+								}
+								
 								
 								
 								
@@ -165,12 +203,12 @@
 					<?php
 						
 							echo '<td class="actions">';
-							
-							echo $this->Html->link('<i class="glyphicon glyphicon-print"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'createPdf', $item['Confirmation']['id']), array('escape' => false, 'target' => '_blank'));
-							
-							if(!empty($item['Billing']['hash']))
-								echo $this->Html->link('<i class="glyphicon glyphicon-link"></i>', '/Rechnung/'.$item['Billing']['hash'], array('escape' => false, 'target' => '_blank'));
-							
+							if(strpos($item['Confirmation']['status'], 'custom') === FALSE){
+								echo $this->Html->link('<i class="glyphicon glyphicon-print"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'createPdf', $item['Confirmation']['id']), array('escape' => false, 'target' => '_blank'));
+								
+								if(!empty($item['Billing']['hash']))
+									echo $this->Html->link('<i class="glyphicon glyphicon-link"></i>', '/Rechnung/'.$item['Billing']['hash'], array('escape' => false, 'target' => '_blank'));
+							}
 							echo $this->Html->link('<i class="tableSetting_btn glyphicon glyphicon-cog"></i>', array('admin' => true, 'controller' => 'Billings', 'action' => 'table_setting', $item['Billing']['id']), array('escape' => false));
