+++ app/View/Elements/backend/portlets/Deliveries/tableContent.ctp
@@ -1,7 +1,7 @@
 <?php
 	foreach ($data as $item):
 	
-		if($item['Confirmation']['cart_id'] != 0) {				
+		if(true) {				
 ?>
 				<tr>
 					<td>
@@ -11,9 +11,11 @@
 						<?php if(empty($item['Delivery']['delivery_number'])) {
 							 echo '-'; 
 							} else {
-								
-								echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'view', $item['Delivery']['id']), array('escape' => false));
-								
+								if(strpos($item['Delivery']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'edit_individual', $item['Delivery']['id']), array('escape' => false));
+								} else {
+									echo $this->Html->link('<i class="glyphicon glyphicon-search"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'view', $item['Delivery']['id']), array('escape' => false));
+								}
 								echo '&nbsp;&nbsp;&nbsp;';
 								echo $item['Delivery']['delivery_number'];	
 							
@@ -27,19 +29,21 @@
 							if(empty($item['Confirmation']['customer_id'])) { echo '-'; } else {
 															
 								echo $item['Confirmation']['customer_id'];	
-								echo '&nbsp;';
-								echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
-									 data-toggle="popover"
-									 data-content="';
-									 	echo $item['Address']['organisation'].'<br>';
-										if(!empty($item['Address']['department'])) {
-											echo $item['Address']['department'].'<br>';
-										}
-									 	echo $item['Address']['name'].'<br>'.
-									 '"
-									 data-trigger="hover"
-								
-								></i>';
+								if(!is_null($item['Delivery']['address_id'])) {
+									echo '&nbsp;';
+									echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
+										 data-toggle="popover"
+										 data-content="';
+										 	echo $item['Address']['organisation'].'<br>';
+											if(!empty($item['Address']['department'])) {
+												echo $item['Address']['department'].'<br>';
+											}
+										 	echo $item['Address']['name'].'<br>'.
+										 '"
+										 data-trigger="hover"
+									
+									></i>';
+								}
 							}
 						} else { echo '-'; } 
 						 ?>
@@ -56,10 +60,11 @@
 					<td>
 						<?php 
 						$cartProducts = "";
-						foreach ($item['Cart']['CartProduct'] as $cartProduct) {
-							$cartProducts = $cartProducts . $cartProduct['amount'].'x '. $cartProduct['Information']['Product']['name'] . ' ( '.$cartProduct['Information']['Product']['product_number'].' )<br>';
-						}						
+												
 						if(empty($item['Cart']['count'])) { echo '-'; } else {
+							foreach ($item['Cart']['CartProduct'] as $cartProduct) {
+								$cartProducts = $cartProducts . $cartProduct['amount'].'x '. $cartProduct['Information']['Product']['name'] . ' ( '.$cartProduct['Information']['Product']['product_number'].' )<br>';
+							}
 							echo $item['Cart']['count'];	
 							echo '&nbsp;';
 							echo '<i class="glyphicon glyphicon-info-sign" style="color: lightblue; cursor: pointer"
@@ -78,11 +83,17 @@
 						<?php if(!isset($item['Delivery']['confirmation_number'])) {
 							 echo '-'; 
 							} else {
-								
-								echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
-								data-content="'.$item['Confirmation']['confirmation_number'].'"
-								data-trigger="hover"></i>',
-								array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));							
+								if(strpos($item['Delivery']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									data-content="'.$item['Confirmation']['confirmation_number'].'"
+									data-trigger="hover"></i>',
+									array('admin' => true, 'controller' => 'Confirmations', 'action' => 'edit_individual', $item['Confirmation']['id']), array('escape' => false));							
+								} else {
+									echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
+									data-content="'.$item['Confirmation']['confirmation_number'].'"
+									data-trigger="hover"></i>',
+									array('admin' => true, 'controller' => 'Confirmations', 'action' => 'view', $item['Confirmation']['id']), array('escape' => false));							
+								}
 							}
 						?>
 					</td>
@@ -90,8 +101,16 @@
 						<?php 
 							
 							if(empty($item['Confirmation']['billing_id'])) {
-								echo $this->Html->link('Rechnung', array('controller' => 'Billings', 'action' => 'convert', 'admin' =>'true', $item['Confirmation']['id']),
+								if(strpos($item['Delivery']['status'], 'custom') !== FALSE){
+									echo $this->Html->link('Rechnung', array('controller' => 'Billings', 'action' => 'add_individual', 'admin' =>'true', $item['Confirmation']['id']),
 																array('class' => 'btn btn-default')); 	
+								} else {
+									echo $this->Html->link('Rechnung', array('controller' => 'Billings', 'action' => 'convert', 'admin' =>'true', $item['Confirmation']['id']),
+																array('class' => 'btn btn-default')); 							
+								}
+									
+									
+								
 							} else {
 								echo $this->Html->link('<i class="glyphicon glyphicon-search" data-toggle="popover" 
 								data-content="'.$item['Delivery']['billing_number'].'"
@@ -111,12 +130,12 @@
 						
 							echo '<td class="actions">';
 							
-							
-							echo $this->Html->link('<i class="glyphicon glyphicon-print"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'createPdf', $item['Delivery']['id']), array('escape' => false, 'target' => '_blank'));
-							
-							if(!empty($item['Delivery']['hash']))
-								echo $this->Html->link('<i class="glyphicon glyphicon-link"></i>', '/Lieferung/'.$item['Delivery']['hash'], array('escape' => false, 'target' => '_blank'));
-							
+							if(strpos($item['Confirmation']['status'], 'custom') === FALSE){
+								echo $this->Html->link('<i class="glyphicon glyphicon-print"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'createPdf', $item['Delivery']['id']), array('escape' => false, 'target' => '_blank'));
+								
+								if(!empty($item['Delivery']['hash']))
+									echo $this->Html->link('<i class="glyphicon glyphicon-link"></i>', '/Lieferung/'.$item['Delivery']['hash'], array('escape' => false, 'target' => '_blank'));
+							}
 							echo $this->Html->link('<i class="tableSetting_btn glyphicon glyphicon-cog"></i>', array('admin' => true, 'controller' => 'Deliveries', 'action' => 'table_setting', $item['Delivery']['id']), array('escape' => false));
